Reorganized tests into UmaTests.h, added InvalidMutationTests.h testing suite for out of bounds tests
